Crate ic_crypto_internal_csp

Interface for the cryptographic service provider

Re-exports

pub use crypto_lib::hash;

Modules

api
Top level traits for interacting with the crypto service provider

crypto_lib
Cryptographic schemes

imported_test_utils
Testing utilities

imported_utilities
A collection of utility functions

keygen
Utilities for key generation and key identifier generation

public_key_store
Interfaces for saving and retrieving public keys

secret_key_store
Interfaces for saving and retrieving secret keys

threshold
Threshold signature implementation for the CSP

tls_stub
TLS utilities

types
Cryptographic types

Structs

Csp
Implements the CryptoServiceProvider for an RNG and a SecretKeyStore.

Traits

CryptoServiceProvider
Describes the interface of the crypto service provider (CSP), e.g. for signing and key generation. The Csp struct implements this trait.