Module ic_crypto_internal_csp::api

Top level traits for interacting with the crypto service provider

Modules

tls_errors
TLS error types

Enums

CspThresholdSignError
Errors occuring while performing threshold signature generation

Traits

CspKeyGenerator
A trait that can be used to generate cryptographic key pairs

CspSecretKeyInjector
Trait for injecting secret keys into the CSP's key store for testing and stubbing purposes.

CspSecretKeyStoreChecker
A trait that allows checking the secret key store for the availability of a key.

CspSigner
A trait that can generate and verify public key signatures

CspTlsClientHandshake
A trait that exposes TLS client-side handshaking

CspTlsServerHandshake
A trait that exposes TLS server-side handshaking

DistributedKeyGenerationCspClient
Crypto service provider (CSP) client for distributed key generation

NiDkgCspClient
Crypto service provider (CSP) client for non-interactive distributed key generation (NiDKG).

NodePublicKeyData
A trait that exposes the information about node public keys and key identifiers.

ThresholdSignatureCspClient
Crypto service provider (CSP) client for threshold signatures.