Trait ic_crypto_internal_csp::api::CspKeyGenerator

pub trait CspKeyGenerator {
    fn gen_key_pair(
        &self, 
        alg_id: AlgorithmId
    ) -> Result<(KeyId, CspPublicKey), CryptoError>;

    fn gen_key_pair_with_pop(
        &self, 
        algorithm_id: AlgorithmId
    ) -> Result<(KeyId, CspPublicKey, CspPop), CryptoError>;

    fn gen_tls_key_pair(
        &mut self, 
        node_id: NodeId, 
        not_after: &str
    ) -> X509PublicKeyCert;
}

A trait that can be used to generate cryptographic key pairs

Required methods

fn gen_key_pair(
&self,
alg_id: AlgorithmId
) -> Result<(KeyId, CspPublicKey), CryptoError>

Generate a public/private key pair.

Arguments

alg_id specifies the algorithm to be used

Returns

The key ID and the public key of the keypair

Errors

CryptoError::InvalidArgument if the algorithm is not supported by the trait implementation. (Note: Currently only BLS12-381 and Ed25519 are supported by implementations of this trait)

fn gen_key_pair_with_pop(
&self,
algorithm_id: AlgorithmId
) -> Result<(KeyId, CspPublicKey, CspPop), CryptoError>

Generate a public/private key pair with proof of possession.

Arguments

alg_id specifies the algorithm to be used

Returns

The key ID referring to the secret key, the public key, and the PoP

Errors

CryptoError::InvalidArgument if the algorithm is not supported by the trait implementation. (Note: Currently only BLS12-381 is supported by implementations of this trait)

fn gen_tls_key_pair(
&mut self,
node_id: NodeId,
not_after: &str
) -> X509PublicKeyCert

Generates TLS key material for node with ID node_id.

The secret key is stored in the key store and used to create a self-signed X.509 public key certificate with

a random serial,

the common name of both subject and issuer being the ToString form of the given node_id,

validity starting at the time of calling this method, and

validity ending at not_after, which must be specified according to section 4.1.2.5 in RFC 5280.

Returns the public key certificate.

Panics

Panics if not_after is not specified according to RFC 5280 or if not_after is in the past.

Loading content...