Trait ic_crypto_internal_csp::api::CspSecretKeyInjector

pub trait CspSecretKeyInjector {
    fn insert_secret_key(&mut self, key_id: KeyId, secret_key: CspSecretKey);
}

Trait for injecting secret keys into the CSP's key store for testing and stubbing purposes.

Required methods

fn insert_secret_key(&mut self, key_id: KeyId, secret_key: CspSecretKey)

Inserts the given secret key into the secret key store with the given key ID. If a key with the given key_id exists, it is overwritten.

Arguments

key_id is the key ID under which the key is inserted into the store.

secret_key is a secret key that is inserted into the store.

Panics

This method panics if inserting the key fails.

Loading content...