Trait ic_crypto_internal_csp::api::CspSecretKeyStoreChecker

pub trait CspSecretKeyStoreChecker {
    fn sks_contains(&self, key_id: &KeyId) -> bool;

    fn sks_contains_tls_key(&self, cert: &X509PublicKeyCert) -> bool;
}

A trait that allows checking the secret key store for the availability of a key.

Required methods

fn sks_contains(&self, key_id: &KeyId) -> bool

Checks whether the store contains a key with the given id.

fn sks_contains_tls_key(&self, cert: &X509PublicKeyCert) -> bool

Checks whether the store contains a private key for the given cert.

Loading content...