Trait ic_crypto_internal_csp::api::CspTlsClientHandshake

pub trait CspTlsClientHandshake {
#[must_use]    fn perform_tls_client_handshake<'life0, 'async_trait>(
        &'life0 self, 
        tcp_stream: TcpStream, 
        self_cert: X509PublicKeyCert, 
        trusted_server_cert: X509PublicKeyCert
    ) -> Pin<Box<dyn Future<Output = Result<(TlsStream, X509), CspTlsClientHandshakeError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

A trait that exposes TLS client-side handshaking

Required methods

#[must_use]fn perform_tls_client_handshake<'life0, 'async_trait>(
&'life0 self,
tcp_stream: TcpStream,
self_cert: X509PublicKeyCert,
trusted_server_cert: X509PublicKeyCert
) -> Pin<Box<dyn Future<Output = Result<(TlsStream, X509), CspTlsClientHandshakeError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Transforms a TCP stream into a TLS stream by performing a TLS client handshake. This allows to set up a TLS connection as a client.

The self_cert is used as client certificate for mutual SSL and the corresponding private key must be in the secret key store. The client will only connect to a server that presents the trusted_server_cert in the TLS handshake.

Hostname verification will not be performed during the handshake.

The given tcp_stream is consumed. If an error is returned, the TCP connection is therefore dropped.

The TLS stream is returned in a form that does not allow for extracting the private key corresponding to self_cert or the TLS session keys.

Errors

CspTlsClientHandshakeError::MalformedSelfCertificate if self_cert is malformed.

CspTlsClientHandshakeError::MalformedServerCertificate if trusted_server_cert is malformed.

CspTlsClientHandshakeError::CreateConnectorError if there is a problem configuring the TLS client for performing the handshake.

CspTlsClientHandshakeError::HandshakeError if there is an error during the TLS handshake, or the handshake fails.

CspTlsClientHandshakeError::SecretKeyNotFound if the secret key corresponding to self_cert cannot be found in the secret key store.

CspTlsClientHandshakeError::MalformedSecretKey if the secret key corresponding to self_cert is malformed in the secret key store.

CspTlsClientHandshakeError::WrongSecretKeyType if the secret key corresponding to self_cert has the wrong type in the secret key store.

Loading content...