Trait ic_crypto_internal_csp::api::DistributedKeyGenerationCspClient

pub trait DistributedKeyGenerationCspClient {
    fn dkg_create_ephemeral(
        &self, 
        dkg_id: IDkgId, 
        node_id: &[u8]
    ) -> Result<(CspEncryptionPublicKey, CspPop), DkgCreateEphemeralError>;

    fn dkg_verify_ephemeral(
        &self, 
        dkg_id: IDkgId, 
        node_id: &[u8], 
        key: (CspEncryptionPublicKey, CspPop)
    ) -> Result<(), DkgVerifyEphemeralError>;

    fn dkg_create_dealing(
        &self, 
        dkg_id: IDkgId, 
        threshold: NumberOfNodes, 
        receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>]
    ) -> Result<CspDealing, DkgCreateDealingError>;

    fn dkg_verify_dealing(
        &self, 
        threshold: NumberOfNodes, 
        receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>], 
        csp_dealing: CspDealing
    ) -> Result<(), DkgVerifyDealingError>;

    fn dkg_create_response(
        &self, 
        dkg_id: IDkgId, 
        verified_csp_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)], 
        my_index: NodeIndex
    ) -> Result<CspResponse, DkgCreateResponseError>;

    fn dkg_verify_response(
        &self, 
        dkg_id: IDkgId, 
        verified_csp_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)], 
        receiver_index: NodeIndex, 
        receiver_key: (CspEncryptionPublicKey, CspPop), 
        response: CspResponse
    ) -> Result<(), DkgVerifyResponseError>;

    fn dkg_create_transcript(
        &self, 
        threshold: NumberOfNodes, 
        verified_keys: &[Option<(CspEncryptionPublicKey, CspPop)>], 
        verified_csp_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)], 
        verified_responses: &[Option<CspResponse>]
    ) -> Result<CspDkgTranscript, DkgCreateTranscriptError>;

    fn dkg_load_private_key(
        &self, 
        dkg_id: IDkgId, 
        csp_transcript: CspDkgTranscript
    ) -> Result<(), DkgLoadPrivateKeyError>;

    fn dkg_create_resharing_dealing(
        &self, 
        dkg_id: IDkgId, 
        threshold: NumberOfNodes, 
        resharing_public_coefficients: CspPublicCoefficients, 
        receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>]
    ) -> Result<CspDealing, DkgCreateReshareDealingError>;

    fn dkg_verify_resharing_dealing(
        &self, 
        threshold: NumberOfNodes, 
        receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>], 
        csp_dealing: CspDealing, 
        dealer_index: NodeIndex, 
        resharing_public_coefficients: CspPublicCoefficients
    ) -> Result<(), DkgVerifyReshareDealingError>;

    fn dkg_create_resharing_transcript(
        &self, 
        threshold: NumberOfNodes, 
        receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>], 
        verified_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)], 
        verified_responses: &[Option<CspResponse>], 
        dealer_keys: &[Option<(CspEncryptionPublicKey, CspPop)>], 
        resharing_public_coefficients: CspPublicCoefficients
    ) -> Result<CspDkgTranscript, DkgCreateReshareTranscriptError>;
}

Crypto service provider (CSP) client for distributed key generation

Distributed Key Generation

Motivation

With a standard threshold signature scheme, one entity generates all the signatories' keys and so is in a position to abuse that power. In distributed key generation, multiple actors, called dealers, collaborate to generate the signatories secret key material in a way that no strict subset of dealers has sufficient information to abuse their position.

Usage

The participants are divided into dealers, who contribute to key generation, and receivers, who will be threshold signatories.

Ephemeral key phase:

All participants generate an ephemeral DiffieHellman key and share their ephemeral public keys with the other participants. This key should be used for one DKG and then discarded.

All participants should verify that the ephemeral public keys are plausibly correct and evict any participants whose public keys are invalid.

Dealing phase:

The dealers each individually generate a threshold polynomial and publish it as a 'dealing'. The secret keys for the receivers are encrypted with DiffieHellman and published with the polynomial; only the corresponding receivers can decrypt the secret keys.

All participants verify that the dealings look plausibly correct and reject any dealings that are invalid.

Complaint phase:

Each receiver individually decrypts their secret key in each dealing and verifies that it is correct; if not they generate a complaint against that dealer.

Each participant verifies that the complaints are superficially valid, rejecting any that are not.

Transcript phase

Each participant examines the complaints and verifies the zero knowledge proof that should be included in a complaint. If a complaint is invalid, all complaints by that receiver are disregarded. Of the remaining complaints, any dealings that have been proven to be invalid are rejected. The remaining valid dealings are combined to generate the threshold key. The valid dealings, together with the threshold public key, are published as the transcript.

Individual signatories compute their threshold secret key from the transcript, decrypting their shares using their ephemeral DiffieHellman key. The ephemeral key is then deleted to protect the threshold key from compromise.

Resharing

Motivation

Signatories may cease to exist or be forced to retire. Under these circumstances it may be desirable to find a new set of signatories to replace the old. However it is desirable for such a change to be invisible from the perspective of a user wishing to verify threshold signatures. That is, it is desirable for the public threshold coefficient at zero to remain the same when there is a change of signatories. The resharing scheme achieves this.

Usage

The resharing protocol is the same as the distributed key generation protocol with these modifications:

A threshold key must preexist.

Signatories of the previous threshold key are the dealers during the resharing DKG.

The arguments and calculation differ slightly at some of the steps in the protocol.

On the inclusion of self in otherwise static methods

A CSP client categorizes the offered methods in two categories:

methods that are served non-statically by calling out to a CSP (server) which includes RPC calls (incl. serialization of parameters and deserialization of return values)

methods that are served statically by computing the output locally, that is, without having to make RPC calls, which are comparatively expensive.

When implementing this scheme,

the methods dkg_create_ephemeral, dkg_create_dealing and dkg_load_private_key are computed non-statically

all other methods are computed statically

The methods in this trait that are meant to be served statically are intentionally made non-static (i.e., they take a &self parameter) so that mocking of this trait (for testing purposes) is possible.

Required methods

fn dkg_create_ephemeral(
&self,
dkg_id: IDkgId,
node_id: &[u8]
) -> Result<(CspEncryptionPublicKey, CspPop), DkgCreateEphemeralError>

Generates an ephemeral key used for distributed key generation of threshold keys. Note: Ephemeral keys are NOT threshold keys

Arguments

dkg_id is the identifier for the distributed key being generated.

node_id is a unique identifier for this DKG participant.

Panics

This method MUST panic if it is unable to store the secret key.

Errors

This method SHALL return an error if:

there is an incompatible preexisting key for this entity and DKG. (MalformedSecretKey)

fn dkg_verify_ephemeral(
&self,
dkg_id: IDkgId,
node_id: &[u8],
key: (CspEncryptionPublicKey, CspPop)
) -> Result<(), DkgVerifyEphemeralError>

Verifies that an ephemeral public key and PoP is valid.

Arguments

dkg_id is the identifier for the distributed key being generated.

node_id is a unique identifier for this DKG participant.

(public_key, pop) is a public key and the corresponding proof of possession.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the public key is not well formed. (MalformedPublicKeyWithPopError)

the CspPop is not consistent with all other arguments. (InvalidPopError)

fn dkg_create_dealing(
&self,
dkg_id: IDkgId,
threshold: NumberOfNodes,
receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>]
) -> Result<CspDealing, DkgCreateDealingError>

Generates keys for each eligible receiver.

Preconditions

The caller MUST verify that the PoPs are valid before calling this method. This method will not return an error if a PoP is invalid but such use is dangerous.

Arguments

dkg_id is the identifier for the distributed key being generated.

threshold is the minimum number of nodes required to generate a valid threshold signature.

receiver_keys is a vector with one entry for each receiver, in order. Eligible receivers have their public key in their slot. Ineligible receivers have none.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

any of the public keys is malformed. (MalformedPublicKeyError)

the ephemeral key for the DKG cannot be found. (KeyNotFoundError)

the threshold or number of receiver keys is out of bounds. (SizeError)

fn dkg_verify_dealing(
&self,
threshold: NumberOfNodes,
receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>],
csp_dealing: CspDealing
) -> Result<(), DkgVerifyDealingError>

Verifies the syntactic correctness of the dealing

This checks that the dealing parses and that the public elements are plausible (degree and group size)

Arguments

threshold is the minimum number of nodes required to generate a valid threshold signature.

receiver_keys is a vector with one entry for each receiver, in order. Eligible receivers have their public key in their slot. Ineligible receivers have none.

csp_dealing is the dealing to be verified.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the csp_dealing is malformed. (MalformedDealingError)

the csp_dealing is incompatible with threshold. (InvalidDealingError)

dealing.receiver_data has a different length from receiver_keys. (InvalidDealingError)

csp_dealing SHOULD contain an encrypted secret key share for the i'th receiver if and only if a key for the i'th receiver is present in receiver_keys (i.e. not None); if this is not the case this method SHALL return an error. (InvalidDealingError)

the number of receiver keys is out of bounds. (SizeError)

fn dkg_create_response(
&self,
dkg_id: IDkgId,
verified_csp_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)],
my_index: NodeIndex
) -> Result<CspResponse, DkgCreateResponseError>

Verifies that an encrypted key for this node is valid.

This is called by a receiver to check that the key they have received from a dealer is valid.

Arguments

dkg_id is the identifier for the distributed key being generated.

verified_csp_dealings - for each dealer who has dealt and not been disqualified, their dealing.

my_index is the receiver index for the current node.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

any of the arguments is malformed. Note, however, that the caller MUST verify dealings before creating a response so this SHOULD never happen. (MalformedPublicKeyWithPopError or MalformedDealingError)

any of the dealings does not contain a share for the current node. (MissingShareError)

there is no ephemeral key for this DKG ID. (KeyNotFoundError)

the number of dealings is out of bounds. (SizeError)

fn dkg_verify_response(
&self,
dkg_id: IDkgId,
verified_csp_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)],
receiver_index: NodeIndex,
receiver_key: (CspEncryptionPublicKey, CspPop),
response: CspResponse
) -> Result<(), DkgVerifyResponseError>

Verifies the complaints and other relevant fields in a response.

If all the complaints in the response are upheld, the response is considered valid and any dealers the response makes complaints against SHOULD be disqualified. On the other hand if any of the complaints are invalid, the entire response SHALL be rejected and complaints SHOULD be discarded, even if some of the complaints are valid.

Arguments

dkg_id is the identifier for the distributed key being generated.

verified_csp_dealings is verified syntactically correct dealings.

receiver_index is the index of the receiver who generated the response.

receiver_key is the public key of the receiver who generated the response.

response is the response being verified.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

any of the fields is malformed. (MalformedPublicKeyWithPopError, MalformedDealingError or MalformedResponseError)

one or more complaints in the response is invalid. (InvalidResponseError)

the number of dealings is out of bounds. (SizeError)

fn dkg_create_transcript(
&self,
threshold: NumberOfNodes,
verified_keys: &[Option<(CspEncryptionPublicKey, CspPop)>],
verified_csp_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)],
verified_responses: &[Option<CspResponse>]
) -> Result<CspDkgTranscript, DkgCreateTranscriptError>

Filters out bad dealings.

Arguments

threshold is the minimum number of nodes required to generate a valid threshold signature.

verified_keys is the ephemeral public keys of all eligible receivers.

verified_csp_dealings is syntactically correct dealings.

verified_responses are, primarily, verified complaints about dealings.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

any of the fields is malformed. (MalformedPublicKeyWithPopError, MalformedDealingError, MalformedResponseError)

all dealings are disqualified by verified responses. (InsufficientValidDealingsError)

there are insufficient valid responses. (InsufficientValidResponsesError)

the threshold, number of keys or dealings is out of bounds. (SizeError)

fn dkg_load_private_key(
&self,
dkg_id: IDkgId,
csp_transcript: CspDkgTranscript
) -> Result<(), DkgLoadPrivateKeyError>

Creates a threshold key in the secret key store.

This allows a recipient to proceed as with any threshold key:

The key ID is derived from the dkg_id with: KeyId::from(dkg_id)

The PublicCoefficients are available from the transcript. See the threshold API for how to proceed further.

Note: This causes the ephemeral key to be deleted, if it is still in the secret key store.

Arguments

dkg_id is the identifier for the distributed key being generated.

csp_transcript is a summary of the key generation.

Panics

This MAY panic if the secret key store is unavailable.

Errors

This method SHALL return an error if:

the transcript is malformed. (MalformedTranscriptError)

it is unable to compute the secret key. (InvalidTranscriptError)

it fails to write the secret key to the key store. (SecretKeyStoreWriteError)

fn dkg_create_resharing_dealing(
&self,
dkg_id: IDkgId,
threshold: NumberOfNodes,
resharing_public_coefficients: CspPublicCoefficients,
receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>]
) -> Result<CspDealing, DkgCreateReshareDealingError>

Generates a fresh set of threshold keys from a threshold secret key held by this node.

Preconditions

The dealers in the current DKG must be signatories for the preceding threshold key.

A new set of ephemeral keys should be used for each DKG; the previous ephemeral keys are likely to have been deleted before this method is called.

The caller MUST verify that the PoPs are valid before calling this method. This method will not return an error if a PoP is invalid but such use is dangerous.

Arguments

dkg_id is the new DKG identifier.

threshold is the minimum number of nodes required to generate a valid threshold signature using the new keys; this may differ from the preceding threshold.

receiver_keys is the ephemeral public keys of all eligible receivers.

resharing_public_coefficients are the public coefficients of the previous DKG that is being reshared.

receiver_keys is a vector with one entry for each receiver, in order. Eligible receivers have their public key in their slot. Ineligible receivers have none.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

any of the public keys is malformed. (MalformedPublicKeyError)

the ephemeral key for the DKG cannot be found. (KeyNotFoundError)

the number of public coefficients or receiver keys is out of bounds. (SizeError)

fn dkg_verify_resharing_dealing(
&self,
threshold: NumberOfNodes,
receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>],
csp_dealing: CspDealing,
dealer_index: NodeIndex,
resharing_public_coefficients: CspPublicCoefficients
) -> Result<(), DkgVerifyReshareDealingError>

Verifies the syntactic correctness of a reshared dealing.

This checks that the dealing parses and that the public elements have the correct degree and group size.

Arguments

threshold is the minimum number of nodes required to generate a valid threshold signature using the new keys; this may differ from the preceding threshold.

receiver_keys is a vector with one entry for each receiver, in order. Eligible receivers have their public key in their slot. Ineligible receivers have none.

csp_dealing is the dealing to be verified.

dealer_index is is the current dealer's index in the preceding threshold key.

resharing_public_coefficients are the public coefficients of the previous threshold key.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the csp_dealing is malformed. (MalformedDealingError)

the csp_dealing has a different threshold from the preceding dealing. (InvalidDealingError)

dealing.receiver_data has a different length from receiver_keys. (InvalidDealingError)

csp_dealing SHOULD contain an encrypted secret key share for the i'th receiver if and only if a key for the i'th receiver is present in receiver_keys (i.e. not None); if this is not the case this method SHALL return an error. (InvalidDealingError)

the number of receiver keys is out of bounds. (SizeError)

fn dkg_create_resharing_transcript(
&self,
threshold: NumberOfNodes,
receiver_keys: &[Option<(CspEncryptionPublicKey, CspPop)>],
verified_dealings: &[((CspEncryptionPublicKey, CspPop), CspDealing)],
verified_responses: &[Option<CspResponse>],
dealer_keys: &[Option<(CspEncryptionPublicKey, CspPop)>],
resharing_public_coefficients: CspPublicCoefficients
) -> Result<CspDkgTranscript, DkgCreateReshareTranscriptError>

Filters out bad resharing dealings.

Arguments

threshold is the minimum number of nodes required to generate a valid threshold signature using the new keys; this may differ from the preceding threshold.

receiver_keys is the ephemeral public keys of all eligible receivers. Note that the order of the keys matters and must be preserved for use with the new threshold key.

verified_dealings are syntactically correct dealings.

verified_responses are, primarily, verified complaints about dealings.

dealer_keys is the ephemeral public keys of dealers. Note that the order of the keys matters; the ith entry in this vector should correspond to the ith receiver of the preceding threshold key.

resharing_public_coefficients are the public coefficients of the previous threshold key. They are used to determine the threshold of the previous threshold key being reshared.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

any of the fields is malformed. (MalformedPublicKeyWithPopError, MalformedDealingError, MalformedResponseError)

all dealings are disqualified by verified responses. (InsufficientValidDealingsError)

there are insufficient valid responses; the minimum is 2*threshold-1. (InsufficientValidResponsesError)

the threshold, number of keys or dealings is out of bounds. (SizeError)

Loading content...