Trait ic_crypto_internal_csp::api::NiDkgCspClient

pub trait NiDkgCspClient {
    fn create_forward_secure_key_pair(
        &mut self, 
        algorithm_id: AlgorithmId, 
        node_id: NodeId
    ) -> Result<(CspFsEncryptionPublicKey, CspFsEncryptionPop), CspDkgCreateFsKeyError>;

    fn verify_forward_secure_key(
        &self, 
        algorithm_id: AlgorithmId, 
        public_key: CspFsEncryptionPublicKey, 
        pop: CspFsEncryptionPop, 
        node_id: NodeId
    ) -> Result<(), CspDkgVerifyFsKeyError>;

    fn update_forward_secure_epoch(
        &self, 
        algorithm_id: AlgorithmId, 
        epoch: Epoch
    ) -> Result<(), CspDkgUpdateFsEpochError>;

    fn create_dealing(
        &self, 
        algorithm_id: AlgorithmId, 
        dkg_id: NiDkgId, 
        dealer_index: NodeIndex, 
        threshold: NumberOfNodes, 
        epoch: Epoch, 
        receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>
    ) -> Result<CspNiDkgDealing, CspDkgCreateDealingError>;

    fn create_resharing_dealing(
        &self, 
        algorithm_id: AlgorithmId, 
        dkg_id: NiDkgId, 
        dealer_resharing_index: NodeIndex, 
        threshold: NumberOfNodes, 
        epoch: Epoch, 
        receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>, 
        resharing_public_coefficients: CspPublicCoefficients
    ) -> Result<CspNiDkgDealing, CspDkgCreateReshareDealingError>;

    fn verify_dealing(
        &self, 
        algorithm_id: AlgorithmId, 
        dkg_id: NiDkgId, 
        dealer_index: NodeIndex, 
        threshold: NumberOfNodes, 
        epoch: Epoch, 
        receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>, 
        dealing: CspNiDkgDealing
    ) -> Result<(), CspDkgVerifyDealingError>;

    fn verify_resharing_dealing(
        &self, 
        algorithm_id: AlgorithmId, 
        dkg_id: NiDkgId, 
        dealer_resharing_index: NodeIndex, 
        threshold: NumberOfNodes, 
        epoch: Epoch, 
        receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>, 
        dealing: CspNiDkgDealing, 
        resharing_public_coefficients: CspPublicCoefficients
    ) -> Result<(), CspDkgVerifyReshareDealingError>;

    fn create_transcript(
        &self, 
        algorithm_id: AlgorithmId, 
        threshold: NumberOfNodes, 
        number_of_receivers: NumberOfNodes, 
        csp_dealings: BTreeMap<NodeIndex, CspNiDkgDealing>
    ) -> Result<CspNiDkgTranscript, CspDkgCreateTranscriptError>;

    fn create_resharing_transcript(
        &self, 
        algorithm_id: AlgorithmId, 
        threshold: NumberOfNodes, 
        number_of_receivers: NumberOfNodes, 
        csp_dealings: BTreeMap<NodeIndex, CspNiDkgDealing>, 
        resharing_public_coefficients: CspPublicCoefficients
    ) -> Result<CspNiDkgTranscript, CspDkgCreateReshareTranscriptError>;

    fn load_threshold_signing_key(
        &self, 
        algorithm_id: AlgorithmId, 
        dkg_id: NiDkgId, 
        epoch: Epoch, 
        csp_transcript: CspNiDkgTranscript, 
        receiver_index: NodeIndex
    ) -> Result<(), CspDkgLoadPrivateKeyError>;

    fn retain_threshold_keys_if_present(
        &self, 
        active_keys: BTreeSet<CspPublicCoefficients>
    );
}

Crypto service provider (CSP) client for non-interactive distributed key generation (NiDKG).

Distributed Key Generation

Motivation

With a standard threshold signature scheme, one entity generates all the signatories' keys and so is in a position to abuse that power. In distributed key generation, multiple actors, called dealers, collaborate to generate the signatories secret key material in a way that no strict subset of dealers has sufficient information to abuse their position.

Non-interactive DKG is a design that is easy to use because there is effectively only a single step: Dealers deal shares and can then cease to exist, as far as the protocol is concerned. Receivers likewise do not need to communicate with each other; they can read the dealings written by the dealers and proceed without further ado. Not only is this design convenient, it also enables use cases in which it is hard for participants to coordinate with each other.

Usage

Participants may be dealers, who contribute to key generation, or receivers, who will be threshold signatories, or both.

Dealing

The dealers each individually generate a threshold polynomial and publish it as a 'dealing'. The secret keys for the receivers are encrypted with a forward secure encryption scheme and published with the polynomial; only the corresponding receivers can decrypt the secret keys.

Each dealing should be verified and accepted only if sound.

Receiving

The receivers each read the dealings and compile them into a transcript.

Each receiver can compute their secret threshold key from the transcript and so contribute to threshold signatures.

Resharing

Motivation

Signatories may cease to exist or be forced to retire. Under these circumstances it may be desirable to find a new set of signatories to replace the old. However it is desirable for such a change to be invisible from the perspective of a user wishing to verify threshold signatures. That is, it is desirable for the public threshold coefficient at zero to remain the same when there is a change of signatories. The resharing scheme achieves this.

Usage

The resharing protocol is the same as the distributed key generation protocol with these modifications:

A threshold key must preexist.

Signatories of the previous threshold key are the dealers during the resharing DKG.

The arguments and calculation differ slightly at some of the steps in the protocol.

TODO(CRP-563): Prefix the NiDkg methods with nidkg_.

TODO(CRP-564): Remove the csp_ prefix from argument names.

Required methods

fn create_forward_secure_key_pair(
&mut self,
algorithm_id: AlgorithmId,
node_id: NodeId
) -> Result<(CspFsEncryptionPublicKey, CspFsEncryptionPop), CspDkgCreateFsKeyError>

Generates a forward secure key pair used to encrypt threshold key shares in transmission.

Note: FS keys are NOT threshold keys.

The secret key is stored in the secret key store. It is not returned by the method as that would violate the principle that secret keys never leave the CSP. The public key and the proof of possession are returned. The public key can be used to verify signatures, it also needs to be provided when signing as it is used to retrieve the secret key from the key store.

Arguments

node_id is the identity of the node generating the public key.

Panics

This method MUST panic if it is unable to store the secret key.

Errors

This method is infallible. An error structure is included in the signature only for future use.

fn verify_forward_secure_key(
&self,
algorithm_id: AlgorithmId,
public_key: CspFsEncryptionPublicKey,
pop: CspFsEncryptionPop,
node_id: NodeId
) -> Result<(), CspDkgVerifyFsKeyError>

Verifies that a forward secure public key and PoP is valid.

Arguments

(public_key, pop) is a public key and the corresponding proof of possession.

node_id is the identity of the node that generated the public key.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the public key is not well formed. (MalformedPublicKeyWithPopError)

the CspPop is not consistent with all other arguments. (InvalidPop)

fn update_forward_secure_epoch(
&self,
algorithm_id: AlgorithmId,
epoch: Epoch
) -> Result<(), CspDkgUpdateFsEpochError>

Updates the secret key corresponding to public_key so that it cannot be used at epochs that are smaller than the given epoch.

Arguments

public_key is the public key corresponding to the secret key to be updated.

epoch is the epoch to be deleted, together with all smaller epochs.

Panics

This method MUST panic if it is unable to store the updated secret key.

Errors

This method SHALL return an error if:

the public key is not well formed. (MalformedPublicKeyError) This method SHALL NOT return an error if:

the forward secure epoch is already higher than the epoch provided in the method argument. In this case the secret key is unchanged.

fn create_dealing(
&self,
algorithm_id: AlgorithmId,
dkg_id: NiDkgId,
dealer_index: NodeIndex,
threshold: NumberOfNodes,
epoch: Epoch,
receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>
) -> Result<CspNiDkgDealing, CspDkgCreateDealingError>

Generates keys for each eligible receiver.

Arguments

algorithm_id selects the algorithm suite to use for the scheme.

dkg_id is the identifier for the distributed key being generated.

dealer_index the index associated with the dealer.

threshold is the minimum number of nodes required to generate a valid threshold signature.

epoch is a monotonic increasing counter used to select forward secure keys.

receiver_keys is a map storing one public key for each receiver, indexed by their corresponding NodeIndex.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the algorithm_id is not an NiDkg variant. (UnsupportedAlgorithmId)

the threshold is less than 1 or greater than the number of receivers. (InvalidThresholdError)

the receiver indices are not 0..num_receivers-1 inclusive. (MisnumberedReceiverError)

one of the receivers' public keys is malformed. (MalformedFsPublicKeyError)

the number of receiver keys is unsupported by this machine. (SizeError)

fn create_resharing_dealing(
&self,
algorithm_id: AlgorithmId,
dkg_id: NiDkgId,
dealer_resharing_index: NodeIndex,
threshold: NumberOfNodes,
epoch: Epoch,
receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>,
resharing_public_coefficients: CspPublicCoefficients
) -> Result<CspNiDkgDealing, CspDkgCreateReshareDealingError>

Generates a fresh set of threshold keys from a threshold secret key held by this node.

Preconditions

The dealers in the current DKG must be signatories for the preceding threshold key.

Arguments

algorithm_id selects the algorithm suite to use for the scheme.

dkg_id is the new DKG identifier.

dealer_resharing_index the index associated with the dealer in the resharing committee.

threshold is the minimum number of nodes required to generate a valid threshold signature using the new keys; this may differ from the preceding threshold.

epoch is a monotonic increasing counter used to select forward secure keys.

receiver_keys is a map storing one public key for each receiver, indexed by their corresponding NodeIndex.

resharing_public_coefficients are the public coefficients of the previous DKG that is being reshared.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the algorithm_id is not an NiDkg variant. (UnsupportedAlgorithmId)

the threshold is less than 1 or greater than the number of receivers. (InvalidThresholdError)

the receiver indices are not 0..num_receivers-1 inclusive. (MisnumberedReceiverError)

one of the receiver keys is invalid. (MalformedFsPublicKeyError)

the key to be reshared is not present or malformed. (ReshareKeyNotFoundError, MalformedReshareSecretKeyError)

the number of public coefficients or receiver keys is unsupported by this machine. (SizeError)

fn verify_dealing(
&self,
algorithm_id: AlgorithmId,
dkg_id: NiDkgId,
dealer_index: NodeIndex,
threshold: NumberOfNodes,
epoch: Epoch,
receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>,
dealing: CspNiDkgDealing
) -> Result<(), CspDkgVerifyDealingError>

Verifies the correctness of the dealing

This checks that the dealing parses and uses zero knowledge proofs to verify that the shares are correct.

Arguments

algorithm_id selects the algorithm suite to use for the scheme.

dkg_id is a unique identifier for the DKG.

dealer_index the index associated with the dealer.

threshold is the minimum number of nodes required to generate a valid threshold signature.

epoch is a monotonic increasing counter used to select forward secure keys.

receiver_keys is a map storing one public key for each receiver, indexed by their corresponding NodeIndex.

dealing is the dealing to be verified.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the algorithm_id is not an NiDkg variant. (UnsupportedAlgorithmId)

the threshold is less than 1 or greater than the number of receivers. (InvalidThresholdError)

the receiver indices are not 0..num_receivers-1 inclusive. (MisnumberedReceiverError)

one of the receiver keys is invalid. (MalformedFsPublicKeyError)

the dealing is malformed. (MalformedDealingError)

the dealing is incompatible with threshold. (InvalidDealingError)

dealing SHOULD contain an encrypted secret key share for the i'th receiver if and only if a key for the i'th receiver is present in receiver_keys; if this is not the case this method SHALL return an error. (InvalidDealingError)

the number of receiver keys is unsupported by this machine. (SizeError)

fn verify_resharing_dealing(
&self,
algorithm_id: AlgorithmId,
dkg_id: NiDkgId,
dealer_resharing_index: NodeIndex,
threshold: NumberOfNodes,
epoch: Epoch,
receiver_keys: BTreeMap<NodeIndex, CspFsEncryptionPublicKey>,
dealing: CspNiDkgDealing,
resharing_public_coefficients: CspPublicCoefficients
) -> Result<(), CspDkgVerifyReshareDealingError>

Verifies the correctness of a reshared dealing.

This checks that the public key is unchanged by resharing as well as all the checks performed by verify_dealing to ensure that the shares are correct.

Arguments

algorithm_id selects the algorithm suite to use for the scheme.

dkg_id is the new DKG identifier.

dealer_resharing_index the index associated with the dealer in the resharing committee.

threshold is the minimum number of nodes required to generate a valid threshold signature using the new keys; this may differ from the preceding threshold.

epoch is a monotonic increasing counter used to select forward secure keys.

receiver_keys is a map storing one public key for each receiver, indexed by their corresponding NodeIndex.

dealing is the dealing to be verified.

resharing_public_coefficients are the public coefficients of the previous threshold key.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the algorithm_id is not an NiDkg variant. (UnsupportedAlgorithmId)

the threshold is less than 1 or greater than the number of receivers. (InvalidThresholdError)

the receiver indices are not 0..num_receivers-1 inclusive. (MisnumberedReceiverError)

one of the receiver keys is invalid. (MalformedFsPublicKeyError)

the reshare public coefficients are malformed. (MalformedResharePublicCoefficientsError)

the dealing is malformed. (MalformedDealingError)

the dealing does not match the threshold. (InvalidDealingError)

dealing SHOULD contain an encrypted secret key share for the i'th receiver if and only if a key for the i'th receiver is present in receiver_keys; if this is not the case this method SHALL return an error. (InvalidDealingError)

the dealing does not reshare the preexisting dealer's public key. (InvalidDealingError)

the number of receiver keys is unsupported by this machine. (SizeError)

fn create_transcript(
&self,
algorithm_id: AlgorithmId,
threshold: NumberOfNodes,
number_of_receivers: NumberOfNodes,
csp_dealings: BTreeMap<NodeIndex, CspNiDkgDealing>
) -> Result<CspNiDkgTranscript, CspDkgCreateTranscriptError>

Assembles all valid dealings into a single record.

Preconditions

The dealings MUST be verified before inclusion.

Arguments

algorithm_id selects the algorithm suite to use for the scheme.

threshold is the minimum number of nodes required to generate a valid threshold signature.

number_of_receivers is the number of receivers that should have shares in each dealing.

csp_dealings are correct dealings.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the algorithm_id is not an NiDkg variant. (UnsupportedAlgorithmId)

the threshold is less than 1 or greater than the number of receivers. (InvalidThresholdError)

any of the dealings is invalid. (InvalidDealingError)

any dealing in csp_dealings does not match number_of_receivers. (InvalidDealingError)

there are insufficient dealings to proceed safely. (InsufficientDataError)

the threshold, number of keys or dealings is unsupported by this machine. (SizeError)

fn create_resharing_transcript(
&self,
algorithm_id: AlgorithmId,
threshold: NumberOfNodes,
number_of_receivers: NumberOfNodes,
csp_dealings: BTreeMap<NodeIndex, CspNiDkgDealing>,
resharing_public_coefficients: CspPublicCoefficients
) -> Result<CspNiDkgTranscript, CspDkgCreateReshareTranscriptError>

Assembles all valid resharing dealings into a single record.

Preconditions

The dealings MUST be verified before inclusion.

Arguments

algorithm_id selects the algorithm suite to use for the scheme.

threshold is the minimum number of nodes required to generate a valid threshold signature using the new keys; this may differ from the preceding threshold.

number_of_receivers is the number of receivers that should have shares in each dealing.

csp_dealings are correct dealings.

resharing_public_coefficients are the public coefficients of the previous threshold key. They are used to determine the threshold and public key of the previous threshold key being reshared.

Panics

This method is not expected to panic.

Errors

This method SHALL return an error if:

the algorithm_id is not an NiDkg variant. (UnsupportedAlgorithmId)

the threshold is less than 1 or greater than the number of receivers. (InvalidThresholdError)

the public key to be reshared is invalid. (MalformedResharePublicCoefficientsError)

any dealing in csp_dealings does not match number_of_receivers. (InvalidDealingError)

there are insufficient dealings to proceed safely. (InsufficientDataError)

the reshared public key changes. This will not happen if all dealings are verified correctly. (ResharingFailed)

the threshold, number of keys or dealings is unsupported by this machine. (SizeError)

fn load_threshold_signing_key(
&self,
algorithm_id: AlgorithmId,
dkg_id: NiDkgId,
epoch: Epoch,
csp_transcript: CspNiDkgTranscript,
receiver_index: NodeIndex
) -> Result<(), CspDkgLoadPrivateKeyError>

Computes a threshold signing key and stores it in the secret key store.

After calling this method to store the key in the secret key store, the threshold signature API can be used to generate signature shares.

Notes:

The key ID is derived from the public coefficients with: KeyId::from(&public_coefficients)

See the threshold API for how to proceed further.

Arguments

algorithm_id selects the algorithm suite to use for the scheme.

dkg_id is the identifier for the distributed key being generated.

epoch is a monotonic increasing counter used to select forward secure keys.

csp_transcript is a summary of the key generation.

public_key is the public part of the forward secure key used to decrypt shares.

receiver_index is the index of the current node in the list of receivers.

Panics

This MAY panic if the secret key store is unavailable.

Errors

This method SHALL return an error if:

the algorithm_id is not an NiDkg variant. (UnsupportedAlgorithmId)

the forward secure decryption key was not found. (KeyNotFoundError)

the forward secure decryption key is malformed. (MalformedSecretKeyError)

the transcript is malformed. (MalformedTranscriptError)

it is unable to compute the secret key. (InvalidTranscriptError)

fn retain_threshold_keys_if_present(
&self,
active_keys: BTreeSet<CspPublicCoefficients>
)

Keeps the specified NiDKG threshold keys.

A threshold key in the secret key store with public coefficients matching an entry in the list will be kept; other threshold keys will be deleted.

There is no guarantee that there are secret keys matching all the listed public coefficients. If this method is requested to retain a key that is not in the secret key store, that key will be ignored.

Loading content...