Trait ic_crypto_internal_csp::api::ThresholdSignatureCspClient

pub trait ThresholdSignatureCspClient {
    fn threshold_sign_to_be_removed(
        &self, 
        algorithm_id: AlgorithmId, 
        message: &[u8], 
        key_id: KeyId
    ) -> Result<CspSignature, CspThresholdSignError>;

    fn threshold_sign(
        &self, 
        algorithm_id: AlgorithmId, 
        message: &[u8], 
        public_coefficients: CspPublicCoefficients
    ) -> Result<CspSignature, CspThresholdSignError>;

    fn threshold_combine_signatures(
        &self, 
        algorithm_id: AlgorithmId, 
        signatures: &[Option<CspSignature>], 
        public_coefficients: CspPublicCoefficients
    ) -> CryptoResult<CspSignature>;

    fn threshold_individual_public_key(
        &self, 
        algorithm_id: AlgorithmId, 
        node_index: NodeIndex, 
        public_coefficients: CspPublicCoefficients
    ) -> CryptoResult<CspThresholdSigPublicKey>;

    fn threshold_verify_individual_signature(
        &self, 
        algorithm_id: AlgorithmId, 
        message: &[u8], 
        signature: CspSignature, 
        public_key: CspThresholdSigPublicKey
    ) -> CryptoResult<()>;

    fn threshold_verify_combined_signature(
        &self, 
        algorithm_id: AlgorithmId, 
        message: &[u8], 
        signature: CspSignature, 
        public_coefficients: CspPublicCoefficients
    ) -> CryptoResult<()>;
}

Crypto service provider (CSP) client for threshold signatures.

A CSP client categorizes the offered methods in two categories:

methods that are served non-statically by calling out to a CSP (server) which includes RPC calls (incl. serialization of parameters and deserialization of return values)

methods that are served statically by computing the output locally, that is, without having to make RPC calls, which are comparatively expensive.

When implementing this client,

the methods for key generation and signing are computed non-statically

all other methods (that is, combining and verifying signatures, and computing public keys) are computed statically

The methods in this trait that are meant to be served statically are intentionally made non-static (i.e., they take a &self parameter) so that mocking of this trait (for testing purposes) is possible.

Required methods

fn threshold_sign_to_be_removed(
&self,
algorithm_id: AlgorithmId,
message: &[u8],
key_id: KeyId
) -> Result<CspSignature, CspThresholdSignError>

Signs with a threshold key

fn threshold_sign(
&self,
algorithm_id: AlgorithmId,
message: &[u8],
public_coefficients: CspPublicCoefficients
) -> Result<CspSignature, CspThresholdSignError>

Signs with a threshold key

fn threshold_combine_signatures(
&self,
algorithm_id: AlgorithmId,
signatures: &[Option<CspSignature>],
public_coefficients: CspPublicCoefficients
) -> CryptoResult<CspSignature>

Combine multiple individual signatures.

Note: This will not check whether the individual signatures are valid, however if all the input signatures are valid and sufficient in number the result will be a valid combined signature.

Note: Signatures must be given in the correct order. A signature by the i'th node must be in the i'th place of the vector of signatures. The vector of signatures may have holes where signatories have not contributed; that is to be expected.

fn threshold_individual_public_key(
&self,
algorithm_id: AlgorithmId,
node_index: NodeIndex,
public_coefficients: CspPublicCoefficients
) -> CryptoResult<CspThresholdSigPublicKey>

Gets an individual signatory's public key.

Note: This is an expensive operation. It is worth keeping the result if it is likely to be reused.

fn threshold_verify_individual_signature(
&self,
algorithm_id: AlgorithmId,
message: &[u8],
signature: CspSignature,
public_key: CspThresholdSigPublicKey
) -> CryptoResult<()>

Checks whether an individual node's signature is valid.

fn threshold_verify_combined_signature(
&self,
algorithm_id: AlgorithmId,
message: &[u8],
signature: CspSignature,
public_coefficients: CspPublicCoefficients
) -> CryptoResult<()>

Checks whether a combined signature is valid. If sufficient valid signatures are combined, the result will pass this test and this is the ultimate goal of the threshold signature scheme.

Loading content...