Module ic_crypto_internal_csp::imported_utilities::sign_utils

Signature utilities

Enums

KeyBytesContentType
Indicates the content type of serialised key bytes passed for parsing.

Functions

combined_threshold_signature_and_public_key
Creates a combined threshold signature together with its public key. This is only used for testing.

ecdsa_p256_signature_from_der_bytes
Decodes an ECDSA P-256 signature from DER.

ed25519_public_key_to_der
Encodes a raw ed25519 public key into DER.

threshold_sig_public_key_from_der
Decodes a threshold signature public key from DER.

threshold_sig_public_key_to_der
Encodes a threshold signature public key into DER.

user_public_key_from_bytes
Parses the given bytes as a DER- or COSE-encoded public key, and returns, if the parsing is successful, the key as UserPublicKey-struct and an enum that indicates the content type of the passed bytes. If parsing fails, returns an error.

verify_combined_threshold_sig
Verifies a combined threshold signature.