Module ic_crypto_internal_csp::keygen

Utilities for key generation and key identifier generation

Modules

utils
Some key related utils

Functions

forward_secure_key_id
Compute the key identifier for a forward secure encryption public key

public_key_hash_as_key_id
Compute the key identifier of the given public key

tls_cert_hash_as_key_id
Create a key identifier by hashing the bytes of the certificate

tls_registry_cert_hash_as_key_id
Create a key identifier by hashing the bytes of the certificate