Trait ic_crypto_internal_csp::secret_key_store::SecretKeyStore

pub trait SecretKeyStore: Send + Sync {
    fn insert(
        &mut self, 
        id: KeyId, 
        key: CspSecretKey, 
        scope: Option<Scope>
    ) -> Result<(), SecretKeyStoreError>;

    fn get(&self, id: &KeyId) -> Option<CspSecretKey>;

    fn contains(&self, id: &KeyId) -> bool;

    fn remove(&mut self, id: &KeyId) -> bool;

    fn insert_or_replace(
        &mut self, 
        id: KeyId, 
        key: CspSecretKey, 
        scope: Option<Scope>
    ) -> Result<(), SecretKeyStoreError> { ... }

    fn retain<F>(&mut self, _filter: F, _scope: Scope)
    where
        F: Fn(&KeyId, &CspSecretKey) -> bool,
    { ... }
}

A store for secret key material

If errors occur regarding reading from or writing to the underlying persistency layer, the methods panic.

Required methods

fn insert(
&mut self,
id: KeyId,
key: CspSecretKey,
scope: Option<Scope>
) -> Result<(), SecretKeyStoreError>

Adds a key with a given id to the store.

Existing keys are not replaced: if a key with the given id already exists, a DuplicateKeyId error is returned. To replace a key with a given id, it must first be removed.

fn get(&self, id: &KeyId) -> Option<CspSecretKey>

Retrieves the key with the given id.

Returns None if the store does not contain a key with the given id.

fn contains(&self, id: &KeyId) -> bool

Checks if the store contains a key with the given id.

fn remove(&mut self, id: &KeyId) -> bool

Removes the key with the given id from the store.

The return value indicates whether a key with the given id was previously contained and removed.

Loading content...
Provided methods

fn insert_or_replace(
&mut self,
id: KeyId,
key: CspSecretKey,
scope: Option<Scope>
) -> Result<(), SecretKeyStoreError>

Inserts a key with a given id into the store, replacing any existing entry.

If there is no existing key with the given id in the store, the new key is inserted.

If there is a key with the given id in the store, it is replaced.

fn retain<F>(&mut self, _filter: F, _scope: Scope) where
F: Fn(&KeyId, &CspSecretKey) -> bool,

Keeps only entries in a scope for which the filter function returns true and removes the rest.

Conversely, this removes entries in a given scope for which the filter function returns false.

Logs

Implementations SHOULD log the KeyId of any deleted keys

Panics

This MAY panic if the predicate panics.

Details: Unlike e.g. HashMap::filter_drain(..) this does not (handle the case where the filter panics)[https://doc.rust-lang.org/std/collections/struct.HashMap.html#method.drain_filter]. If the filter panics, the panic is likely to cause the secret key store to crash. Thus lambdas MUST NEVER be taken from untrusted sources. If panics are to be handled the predicate can be run in a separate thread and panics handles with (thread::Result)[https://doc.rust-lang.org/std/thread/type.Result.html] (depending on whether we are run from inside a suitable runtime), or (panic::catch_unwind(..))[https://doc.rust-lang.org/nightly/std/panic/fn.catch_unwind.html] can be added to this implementation and we may require panic="unwind". See the (book)[https://doc.rust-lang.org/edition-guide/rust-2018/error-handling-and-panics/controlling-panics-with-std-panic.html] and function documentation for more details.

Loading content...