Struct ic_crypto_internal_csp::Csp

pub struct Csp<R: Rng + CryptoRng, S: SecretKeyStore> { /* fields omitted */ }

Implements the CryptoServiceProvider for an RNG and a SecretKeyStore.

Implementations

impl Csp<OsRng, ProtoSecretKeyStore>

pub fn new(
config: &CryptoConfig,
logger: Option<ReplicaLogger>,
metrics: Option<Metrics>
) -> Self

Creates a production-grade crypto service provider.

impl<R: Rng + CryptoRng> Csp<R, ProtoSecretKeyStore>

pub fn new_with_rng(csprng: R, config: &CryptoConfig) -> Self

Creates a crypto service provider for testing.

Note: This MUST NOT be used in production as the secrecy of the random number generator, hence the keys, is not guaranteed.

impl<R: Rng + CryptoRng> Csp<R, VolatileSecretKeyStore>

pub fn reset_public_key_data(&mut self, node_public_keys: NodePublicKeys)

Resets public key data according to the given NodePublicKeys.

Note: This is for testing only and MUST NOT be used in production.

impl<R: Rng + CryptoRng, S: SecretKeyStore> Csp<R, S>

pub fn of(csprng: R, secret_key_store: S) -> Self

Creates a crypto service provider for testing.

Note: This MUST NOT be used in production as the secrecy of the secret key store is not guaranteed.