Module ic_crypto_internal_csp::threshold::dkg

Distributed key generation

This file implements the CSP distributed key generation API. The calculation is done by passing calls through to the crypto lib. The code here:

Converts the arguments into the type accepted by the crypto lib.

Inserts or retrieves secret keys from the secret key store.

Functions

public_coefficients_key_id
Compute a key identifier from public coefficients