Module ic_crypto_internal_csp::threshold::ni_dkg::specialise

The API methods sometimes take multiple arguments, but those arguments must be of compatible types.

Potential strategies for dealing with the potential error of incompatible types:

Give the caller the problem of using the correct types:

Tag types with the scheme they are to be used with. E.g. FsEncryptionPublicKey.

Con: It prevents one type from being used in multiple schemes. Perhaps the caller could do type conversion and we would have alist of legal conversions.

Make every method take a single argument; an enum of structs, each struct corresponding to a scheme.

Con: If the caller is to be responsible for ensuring that types are correct, the caller needs to be aware of the difference between different schemes, however it is an explicit design goal that the IDKM and above should be unaware of algorithm changes. It seems that the caller cannot be made responsible for providing the correct type variants.

Deal with the type conversion ourselves:

Pro: Simplifies life for the caller

Con: Makes no real attempt to design errors out of existence, if that is possible in this case.

We choose to deal with the type conversion internally. These are the type conversions, or to be more specific, type specialisations:

Modules

groth20
Type specialisations for the NiDkg_Groth20 scheme.

Structs

SpecialisationError
An error during specialisation

Functions

fs_key_set
Converts a secret key into a forward secure secret key set.