Module ic_crypto_internal_csp::threshold::ni_dkg::specialise::groth20

Type specialisations for the NiDkg_Groth20 scheme.

Functions

dealing
Converts a generic dealing to the variant used by NiDkg_Groth20.

dealings
Converts a map of generic dealings to the variant used by NiDkg_Groth20.

fs_key_set
Converts a secret key set to the variant used by NiDkg_Groth20.

fs_pop
Converts a proof of possession for a forward secure key to the variant used by NiDkg_Groth20.

fs_public_key
Converts a forward secure public key to the variant used by NiDkg_Groth20.

public_coefficients
Note: We use the CSP equivalent of the g20_internal_types::PublicCoefficientsBytes.

receiver_keys
Converts a map of generic public keys to the variant used by NiDkg_Groth20.

threshold_secret_key
Converts a threshold secret key to the variant used by NiDkg_Groth20.

transcript
Note: We use the CSP equivalent of the g20_internal_types::PublicCoefficientsBytes.