Function ic_crypto_internal_csp::threshold::ni_dkg::specialise::groth20::receiver_keys

pub fn receiver_keys(
    receiver_keys: &BTreeMap<NodeIndex, CspFsEncryptionPublicKey>
) -> Result<BTreeMap<NodeIndex, FsEncryptionPublicKey>, (NodeIndex, MalformedPublicKeyError)>

Converts a map of generic public keys to the variant used by NiDkg_Groth20.

Errors

This returns an error if at least one of the public keys is not of the type expected by NiDkg_Groth20.